decompose shared ec2/gha programming into reusable modular components
